Updated getter to check if item exists as a key
If it does not, it will instead return the parameter. This is to error proof the code. When an unique item name comes up that isn't in our list it will simply spit that same thing onto the table. This way, while updating, I can determine if a recipe is correct, which is a problem I ran into while updating saddler.js.